removed useless item classes
