--- src/main/java/com/example/dreamwishes/dto/ItemsDTO.java
@@ -1,44 +1,0 @@
-package com.example.dreamwishes.dto;
-
-
-public class ItemsDTO {
-    private Long id;
-    private String name;
-    private String description;
-    private double price;
-
-    // Getters and setters
-    // You can also include additional constructors and methods as needed
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-}
