+++ src/main/java/com/example/dreamwishes/dto/WishlistDTO.java
@@ -1,14 +1,12 @@
 package com.example.dreamwishes.dto;
 
 
-import com.example.dreamwishes.model.WishesModel;
-
-import java.util.List;
+import com.example.dreamwishes.model.WishlistModel;
 
 public class WishlistDTO {
     private Long id;
     private Long userId;
-    private List<ItemsDTO> items;
+
     // Getters and setters
 
     public Long getId() {
@@ -27,15 +25,7 @@ public class WishlistDTO {
         this.userId = userId;
     }
 
-    public List<ItemsDTO> getItems() {
-        return items;
-    }
-
-    public void setItems(List<ItemsDTO> items) {
-        this.items = items;
-    }
-
-    public void addWish(WishesModel wishModel) {
+    public void addWish(WishlistModel wishModel) {
     }
