--- src/main/java/com/example/dreamwishes/service/DreamWishesService.java
@@ -1,31 +1,0 @@
-package com.example.dreamwishes.service;
-
-/*
-@Service
-public class DreamWishesService {
-
-    private final WishesRepository wishesRepository;
-
-    @Autowired
-    public DreamWishesService(WishesRepository wishesRepository) {
-        this.wishesRepository = wishesRepository;
-    }
-
-
-    public List<Wishes> getAllWishes() {
-        return wishesRepository.findAll();
-    }
-
-    public Optional<Wishes> getAttractionByName(String name) {
-        return wishesRepository.findByName(name);
-    }
-
-    public void saveAttraction(Wishes attraction) {
-        wishesRepository.save(attraction);
-    }
-    public void deleteAttractionByName(String name) {
-        wishesRepository.deleteAttractionByName(name);
-    }
-}
-
- */
