+++ src/main/resources/templates/index.html
@@ -58,7 +58,8 @@
         </li>
     </ul>
     <ul>
-        <li><a href="/api/wishlists/add">Add an Item</a></li>
+        <li><a href="/api/wishlists/add">Add a wish</a></li>
+        <li><a href="/items/add">Add an Item</a></li>
         <li><a href="/wishlists/user">Go to Wish List</a></li>
     </ul>
 </div>
