+++ src/main/resources/templates/wishes.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>User Wishlist</title>
-  <link rel="stylesheet" type="text/css" href="styles.css">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>User Wishlist</title>
+    <link rel="stylesheet" type="text/css" href="styles.css">
 </head>
 <body>
 
@@ -12,51 +12,52 @@
 
 <!-- Display user information -->
 <div>
-  <h2>User Information:</h2>
-  <p>User ID: <span th:text="${user.id}"></span></p>
-  <p>Username: <span th:text="${user.username}"></span></p>
-  <!-- Add more user information as needed -->
+    <h2>User Information:</h2>
+    <p>User ID: <span th:text="${user.id}"></span></p>
+    <p>Username: <span th:text="${user.username}"></span></p>
+    <!-- Add more user information as needed -->
 </div>
 
 <!-- Display wishlist items -->
 <div>
-  <h2>Wishlist Items:</h2>
-  <ul>
-    <!-- Loop through each item in the wishlist -->
-    <li th:each="item : ${wishlist.item}"> <!-- Changed from ${wishlist.items} to ${wishlist.item} -->
-      <!-- Display item details -->
-      <p>Item ID: <span th:text="${item.id}"></span></p>
-      <p>Name: <span th:text="${item.name}"></span></p>
-      <p>Description: <span th:text="${item.description}"></span></p>
-      <!-- Add more item details as needed -->
+    <h2>Wishlist Items:</h2>
+    <ul th:each="wishlist : ${wishlists}">
+        <!-- Display item details -->
+        <li>
+            <p>Item Name: <span th:text="${wishlist.itemName}"></span></p>
+            <p>Description: <span th:text="${wishlist.description}"></span></p>
+            <p>Price: <span th:text="${wishlist.price}"></span></p>
+            <p>Available: <span th:text="${wishlist.available ? 'Yes' : 'No'}"></span></p>
+            <p>Category: <span th:text="${wishlist.category}"></span></p>
+            <!-- Add more item details as needed -->
 
-      <!-- Add buttons for actions -->
-      <form th:action="@{/wishlist/delete}" method="post">
-        <input type="hidden" name="itemId" th:value="${item.id}">
-        <input type="submit" value="Delete">
-      </form>
+            <!-- Add buttons for actions -->
+            <form th:action="@{/wishlist/delete}" method="post">
+                <input type="hidden" name="wishlistId" th:value="${wishlist.wishlistId}">
+                <input type="submit" value="Delete">
+            </form>
 
-      <!-- Add a form for updating item details -->
-      <form th:action="@{/wishlist/update}" method="post">
-        <input type="hidden" name="itemId" th:value="${item.id}">
-        <input type="text" name="name" placeholder="New name" required>
-        <input type="text" name="description" placeholder="New description" required>
-        <!-- Add more fields for updating item details as needed -->
-        <input type="submit" value="Update">
-      </form>
-    </li>
-  </ul>
+            <!-- Add a form for updating item details -->
+            <form th:action="@{/wishlist/update}" method="post">
+                <input type="hidden" name="wishlistId" th:value="${wishlist.wishlistId}">
+                <input type="text" name="itemName" placeholder="New name" required>
+                <input type="text" name="description" placeholder="New description" required>
+                <!-- Add more fields for updating item details as needed -->
+                <input type="submit" value="Update">
+            </form>
+        </li>
+    </ul>
 </div>
 
 <!-- Form for adding a new item to the wishlist -->
 <div>
-  <h2>Add New Item to Wishlist:</h2>
-  <form th:action="@{/wishlist/add}" method="post">
-    <input type="text" name="name" placeholder="Name" required>
-    <input type="text" name="description" placeholder="Description" required>
-    <!-- Add more fields for adding a new item as needed -->
-    <input type="submit" value="Add">
-  </form>
+    <h2>Add New Item to Wishlist:</h2>
+    <form th:action="@{/wishlist/add}" method="post">
+        <input type="text" name="itemName" placeholder="Name" required>
+        <input type="text" name="description" placeholder="Description" required>
+        <!-- Add more fields for adding a new item as needed -->
+        <input type="submit" value="Add">
+    </form>
 </div>
 
 </body>
